Calculate error
Code to calculate error in Q and eta using the bootstrap method using the result from the code New_plot.sh and can be customised according to the number of iterations and will generate a text file containing the error. Only run bootStrap.sh. It will run the other two codes automatically